// src/components/templates/auth/AuthOptions.types.ts
/** @format */

import { MouseEventHandler } from "react";
export interface AuthOptionsProps {
  text?: string;
  variant?: "register" | "login";
  loading?: boolean;
  onClick?: MouseEventHandler<HTMLButtonElement>;
  formikRegisterStore?: Function;
  formikLoginStore?: Function;
  handleSubmitRegister?: Function;
  handleSubmitLogin?: Function;
}